+++ dzgui/views/base.py
@@ -8,6 +8,7 @@ import signal
 import subprocess
 import textwrap
 import threading
+import traceback
 import typing  # noqa
 import warnings